Make pin login a system endpoint
This endpoint is called by apigw, and never uses the
AuthenticatedUser.MobileDevice parameter in any way
